requirements.txt
torch
torchvision
matplotlib
numpy
tqdm
argparse
Pillow
torch_dct